requirements.txt
ollama==0.5.1
RPi.GPIO==0.7.1
